Added share_utils directory, refactored utilities
