+++ addons/dvan/share_utils/legend_entry.js
