+++ addons/dvan/share_utils/tree_entry.js
